Added Dear Imgui file selection window
Added dear imgui to the project. Added a window that lists the files in the images/ folder, allows the user to select a file which is then rendered to the main window at the appropriate resolution. Added ImageBrowser class that handles the selection of files and displaying the images/ folder.