Added BMP file support
Added support for BMP files with 24 bits per pixel and no compression. Updated ImageRender class to accomidate new file type, as well as the option to flip an image in the fragment shader. Updated the ppm_parser to use a list of char's in memory instead of a vector. Added some sample bmp files. Updated the Shader class to use inbuilt shader paths instead of taking them as arguments at creation.