Added ImageInfo class
Added a new ImageInfo class that uses ImGui to display information about the current image. It is a static singleton class that can be accessed anywhere so the image parsers can easily send it information to display. Information is displayed in tabs which hold pairs of strings: key and value. Tabs must be defined with a name and number of items it will hold. Items are displayed in order of their addition to the ImageInfo.items vector. The number of items must add up to the total size of tabs.

Added windows folder to hold files that output to a window. Updated PNGParser to utilise the ImageInfo class to display the file size and IHDR chunk info.